Move playing music/applause into separate functions
Also have a variable to easily turn on or off pyglet tick polling which
seemed to have been done to prevent music skipping. It probably needs to
be reassesed if this is still necesary. For now it is set to True so it
still enables polling as it did before to prevent skipping.