Add debug_msg() and error_msg() to make debugging easier
Running the program with --debug will enable debug messages, and after
these are used we can print out when an exeption happens and hopefully
be able to debug users with issues much more easily.
Brain Workshop currently relies on exceptions in many cases when it
probably shouldn't, or at minimum should not catch any exception which
occurs.

Adding these I should be able to run the program with --debug and figure
out which exceptions are occuring in normal operation and change the
code so they don't have to occur. We want to get to where we only use
them to catch unexpected things and if those happen to print out a
message even when we aren't on debug mode.