Add LICENSE file and clarify GPL-2.0 or later
Previously there was no LICENSE file, but the header of brainworkshop.py
specified the project was under the GPL. The GPL itself says that if no
version is specified, it is allowed to use any GPL version.

To better clarify the license of the project for the users, as well as
to ensure the license will show up clearly on Github, xantares and I
have decided to officially give it GPL-2.0 or later.